[BatchMatmul] Refactor kernel functions (#538)
* Refactor kernel fns

* format

* Fix includes
