+++ include/flexflow/ops/batch_matmul.h
@@ -5,12 +5,6 @@
 
 namespace FlexFlow {
 
-class BatchMatmulMeta : public OpMeta {
-public:
-  BatchMatmulMeta(FFHandler handler);
-  int a_seq_length_dim, b_seq_length_dim;
-};
-
 class BatchMatmul : public Op {
 public:
   using Params = BatchMatmulParams;
@@ -56,56 +50,6 @@ public:
                             std::vector<Legion::PhysicalRegion> const &regions,
                             Legion::Context ctx,
                             Legion::Runtime *runtime);
-  static void forward_kernel(BatchMatmulMeta const *meta,
-                             float *o_ptr,
-                             float const *a_ptr,
-                             float const *b_ptr,
-                             float const *c_ptr,
-                             int m,
-                             int n,
-                             int k,
-                             int batch,
-                             ffStream_t stream,
-                             int a_seq_length_dim = -1,
-                             int b_seq_length_dim = -1,
-                             int seq_length = -1);
-  static void forward_kernel_wrapper(BatchMatmulMeta const *meta,
-                                     float *o_ptr,
-                                     float const *a_ptr,
-                                     float const *b_ptr,
-                                     float const *c_ptr,
-                                     int m,
-                                     int n,
-                                     int k,
-                                     int batch,
-                                     int a_seq_length_dim = -1,
-                                     int b_seq_length_dim = -1,
-                                     int seq_length = -1);
-  static void backward_kernel(BatchMatmulMeta const *meta,
-                              float const *o_ptr,
-                              float const *o_grad_ptr,
-                              float const *a_ptr,
-                              float *a_grad_ptr,
-                              float const *b_ptr,
-                              float *b_grad_ptr,
-                              float *c_grad_ptr,
-                              int m,
-                              int n,
-                              int k,
-                              int batch,
-                              ffStream_t stream);
-  static void backward_kernel_wrapper(BatchMatmulMeta const *meta,
-                                      float const *o_ptr,
-                                      float const *o_grad_ptr,
-                                      float const *a_ptr,
-                                      float *a_grad_ptr,
-                                      float const *b_ptr,
-                                      float *b_grad_ptr,
-                                      float *c_grad_ptr,
-                                      int m,
-                                      int n,
-                                      int k,
-                                      int batch);
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
